Trying to add AVAudioFile as 3rd reverser
AVAudioFile can do random-access, which was the deal-killer with AVAssetReader/Writer. Only problem at this point is doing the in-memory shuffling of the samples in the read buffer. Seems like directly manipulating that buffer doesn't matter: original buffers get written to the output file.

The ReversePlaybackModel class still uses C as its reversing engine if you actually run the code; change the enum value if you want to play with the broken AVAudioFile stuff. I'm just checking it in for safe-keeping, rather than leave it on my computer until whenever I get back to it.